+++ serve/server.go
@@ -33,8 +33,6 @@ func splitFormValue(formValue string) []string {
 }
 
 type server struct {
-	// https://developer.mozilla.org/en-US/docs/Glossary/Origin
-	HTTPOrigin string
 
 	// Motd - message of the day
 	Motd string
@@ -115,14 +113,9 @@ func (s *server) Submit(c echo.Context) error {
 	// This only works for GET
 	// filledFormURL := s.HTTPOrigin + "/?" + c.Request().URL.RawQuery
 
-	// NOTE: I'm not sure whether the correct HTTP origin can be read from the request. So let's try a few different ways to get it and control them with a flag so I can easily revert in "prod". if I can confirm that "request.Host" works, I can remove the flag and just use c.Request().Host by default. TODO: when I clean this up, also figure out how to get whether the proto is HTTP or HTTPS
-	var httpOrigin string
-	switch s.HTTPOrigin {
-	case "request.Host":
-		httpOrigin = "http://" + c.Request().Host
-	default:
-		httpOrigin = s.HTTPOrigin
-	}
+	// TODO: don't hardcode the http protocol...
+	httpOrigin := "http://" + c.Request().Host
+
 	// This works for POST and I think it works for GET too?
 	filledFormURL := httpOrigin + "/?" + c.Request().Form.Encode()